Fix compilation errors
- Add stdio.h to nic.c for sprintf/sscanf
- Add ipv4.h to transport.h for ipv4_header_t
- Fix shell.c parsing bug with pointer increment
- Add stddef.h to utils.h for size_t
- Fix utils_strlen return type (int vs size_t)
- Fix arp_send_request call (pass value not pointer)

+++ src/shell/utils.c
@@ -202,8 +202,8 @@ char* utils_strncpy(char* dest, const char* src, size_t n) {
 /*
  * Get string length
  */
-size_t utils_strlen(const char* s) {
-    size_t len = 0;
+int utils_strlen(const char* s) {
+    int len = 0;
     while (*s++) len++;
     return len;
 }
